Enable compatibility with EngineFixes. Just a few lines in ScriptExtender.cpp to load the DLLs in the right order.
But, EF isn't compatible if some features are turned on, notoriously tbbmalloc. The bulk of the change is in DllBlocklist, to detect bad configurations of the mod and fix them. It's general-purpose enough to use with other mods if / when that comes up.

Would have been a lot simpler but for TOML configuration rules.