Rewrite of fix for excess logging.
Now, the queue spell/effects is as deep as it needs to be instead of 1 deep. It appears 10s of effects per NPC are needed during server connect or cell load-in.

Order is now preserved, as well as it can be. The code has always sent some out-of-order events before the replay code executes.

Made logging configurable. Default is Debug. Can be different in private build or change it in the debugger. See end of MagicService.h